Module Mindstorm.EV3

type usb

type bluetooth

type 'a conn

Abstract type representing a connection to a LEGO® mindstorm EV3 brick. The type parameter indicates whether this connection is a USB, a WIFI, or a bluetooth one.

val connect_bluetooth : string -> bluetooth conn

connect_bluetooth bdaddr connects through bluetooth to the brick with bluetooth address bdaddr. See the section "connectBluetooth" for more information.

raises Unix.Unix_error
in case of a connection problem. In particular, Unix.Unix_error(Unix.EHOSTDOWN, _,_) is raised if the brick is not turned on.

val close : _ conn -> unit

close conn closes the connection conn to the brick.

module Sound : sig ... end