Module NXT.Program

Starting and stopping programs (.rxe files) on the brick.

val start : ?check_status:bool -> 'a conn -> string -> unit

start_program conn pgm starts the program named pgm.

parameter check_status
whether to check the status returned by the brick. Default: see Mindstorm.NXT.connect_bluetooth.

val stop : ?check_status:bool -> 'a conn -> unit

stop_program conn stops the currently running program if any. If no program is running and check_status=true, the exception Mindstorm.NXT.Error(No_program) is raised.

parameter check_status
whether to check the status returned by the brick. Default: see Mindstorm.NXT.connect_bluetooth.

val name : 'a conn -> string

Return the name of the current program or raise Mindstorm.NXT.Error(No_program) if no program is running.