Module NXT.Sound

Play sounds (.rso files) and tones.

val play : ?check_status:bool -> 'a conn -> ?loop:bool -> string -> unit

play_soundfile conn file plays the sound file named file. The sound files extension, namely ".rso", must be part of file.

parameter loop
if true repeat the play indefinitely. Default: false.

parameter check_status
whether to check the status returned by the brick. Default: see Mindstorm.NXT.connect_bluetooth.

val stop : ?check_status:bool -> 'a conn -> unit

Stop the current playback. Does nothing if no sound file is playing.

parameter check_status
whether to check the status returned by the brick. Default: see Mindstorm.NXT.connect_bluetooth.

val play_tone : ?check_status:bool -> 'a conn -> int -> int -> unit

play_tone conn freq duration play a tone with freq Hz lasting duration miliseconds.

parameter check_status
whether to check the status returned by the brick. Default: see Mindstorm.NXT.connect_bluetooth.